update_09-11-2023

+++ src/main/java/fr/inrae/act/bagap/chloe/window/analysis/sliding/MultipleSlidingLandscapeMetricAnalysis.java
@@ -2,8 +2,11 @@ package fr.inrae.act.bagap.chloe.window.analysis.sliding;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.Arrays;
+import java.util.ArrayList;
 import java.util.HashSet;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
 import java.util.Set;
 
 import fr.inra.sad.bagap.apiland.core.element.manager.Tool;
@@ -24,7 +27,9 @@ public class MultipleSlidingLandscapeMetricAnalysis extends MultipleLandscapeMet
 	
 	private Set<Integer> coherences;
 	
-	private Set<String> csvOutputs;
+	private Map<String, List<String>> csvOutputs;
+	
+	private Map<String, List<String>> suffixMetrics;
 	
 	public MultipleSlidingLandscapeMetricAnalysis(LandscapeMetricAnalysisBuilder builder) {
 		super(builder);
@@ -35,22 +40,82 @@ public class MultipleSlidingLandscapeMetricAnalysis extends MultipleLandscapeMet
 		try {
 			totalMetrics = builder.getMetrics();
 			totalCsvOutput = builder.getCsv();
-			String path = new File(totalCsvOutput).getParent();
-			csvOutputs = new HashSet<String>();
+			
+			String path = null;
+			if(totalCsvOutput != null){
+				path = new File(totalCsvOutput).getParent();
+				csvOutputs = new LinkedHashMap<String, List<String>>();
+				suffixMetrics = new LinkedHashMap<String, List<String>>();
+			}
+			
 			coherences = MetricManager.getCoherences(totalMetrics);
 			Set<Metric> metrics = new HashSet<Metric>();
-			for(int coherence : coherences){
-				metrics.addAll(MetricManager.getMetricsByCoherence(totalMetrics, coherence));
-				if(coherence == 0){
-					continue;
-				}
-				builder.setMetrics(metrics);
-				builder.addCsvOutput(path+"/sliding_"+coherence+".csv");
-				csvOutputs.add(path+"/sliding_"+coherence+".csv");
-				
-				add(LandscapeMetricAnalysisFactory.create(builder));
 			
-				metrics = new HashSet<Metric>();
+			if(builder.getRasterFiles().size() == 0 || builder.getRasterFiles().size() == 1){
+				
+				String rasterFile = builder.getRasterFile();
+				
+				if(totalCsvOutput != null){
+					csvOutputs.put(rasterFile, new ArrayList<String>());
+					suffixMetrics.put(rasterFile, new ArrayList<String>());
+				}
+				
+				for(int ws : builder.getWindowSizes()){
+					builder.setWindowSize(ws);
+					
+					for(int coherence : coherences){
+						metrics.addAll(MetricManager.getMetricsByCoherence(totalMetrics, coherence));
+						if(coherence == 0){
+							continue;
+						}
+						builder.setMetrics(metrics);
+						
+						if(totalCsvOutput != null){
+							builder.addCsvOutput(path+"/sliding_"+coherence+"_"+ws+".csv");
+							csvOutputs.get(rasterFile).add(path+"/sliding_"+coherence+"_"+ws+".csv");
+							suffixMetrics.get(rasterFile).add("_"+ws);
+						}
+						
+						add(LandscapeMetricAnalysisFactory.create(builder));
+						
+						metrics = new HashSet<Metric>();
+					}
+				}
+				
+			}else{
+				
+				for(String rasterFile : builder.getRasterFiles()){
+					String name = new File(rasterFile).getName().replace(".tif", "").replace(".asc", "");
+					builder.setRasterFile(rasterFile);
+					
+					if(totalCsvOutput != null){
+						csvOutputs.put(rasterFile, new ArrayList<String>());
+						suffixMetrics.put(rasterFile, new ArrayList<String>());
+					}
+					
+					for(int ws : builder.getWindowSizes()){
+						builder.setWindowSize(ws);
+						
+						for(int coherence : coherences){
+							metrics.addAll(MetricManager.getMetricsByCoherence(totalMetrics, coherence));
+							if(coherence == 0){
+								continue;
+							}
+							builder.setMetrics(metrics);
+							
+							if(totalCsvOutput != null){
+								builder.addCsvOutput(path+"/"+name+"_sliding_"+coherence+"_"+ws+".csv");
+								csvOutputs.get(rasterFile).add(path+"/"+name+"_sliding_"+coherence+"_"+ws+".csv");
+								suffixMetrics.get(rasterFile).add("_"+ws);
+							}
+							
+							add(LandscapeMetricAnalysisFactory.create(builder));
+						
+							metrics = new HashSet<Metric>();
+						}
+					}
+				}
+				
 			}
 		
 		} catch (IOException e) {
@@ -68,14 +133,50 @@ public class MultipleSlidingLandscapeMetricAnalysis extends MultipleLandscapeMet
 	@Override
 	protected void doClose() {
 		
-		EnteteRaster entete = EnteteRaster.readFromAsciiGridHeader(csvOutputs.iterator().next().replace(".csv", "_header.txt"));
+		if(totalCsvOutput != null){
+			
+			if(builder.getRasterFiles().size() == 0 || builder.getRasterFiles().size() == 1){
+				
+				String rasterFile = builder.getRasterFile();
+					
+				EnteteRaster entete = EnteteRaster.readFromAsciiGridHeader(csvOutputs.get(rasterFile).iterator().next().replace(".csv", "_header.txt"));
+				SpatialCsvManager.mergeXY(totalCsvOutput, csvOutputs.get(rasterFile).toArray(new String[csvOutputs.size()]), suffixMetrics.get(rasterFile).toArray(new String[suffixMetrics.get(rasterFile).size()]), "X", "Y", entete);
+					
+				for(String csvOut : csvOutputs.get(rasterFile)){
+					new File(csvOut).delete();
+					Tool.copy(csvOut.replace(".csv", "_header.txt"), totalCsvOutput.replace(".csv", "_header.txt"));
+					new File(csvOut.replace(".csv", "_header.txt")).delete();
+				}
+				
+			}else{
+				
+				Set<String> localCsv = new HashSet<String>();
+				for(String rasterFile : builder.getRasterFiles()){
+					String name = new File(rasterFile).getName().replace(".tif", "").replace(".asc", "");
+					String localCsvOutput = totalCsvOutput.replace(".csv", "")+"_"+name+".csv";
+					
+					EnteteRaster entete = EnteteRaster.readFromAsciiGridHeader(csvOutputs.get(rasterFile).iterator().next().replace(".csv", "_header.txt"));
+					SpatialCsvManager.mergeXY(localCsvOutput, csvOutputs.get(rasterFile).toArray(new String[csvOutputs.size()]), suffixMetrics.get(rasterFile).toArray(new String[suffixMetrics.get(rasterFile).size()]), "X", "Y", entete);
+					
+					for(String csvOut : csvOutputs.get(rasterFile)){
+						new File(csvOut).delete();
+						Tool.copy(csvOut.replace(".csv", "_header.txt"), localCsvOutput.replace(".csv", "_header.txt"));
+						new File(csvOut.replace(".csv", "_header.txt")).delete();
+					}
+					
+					localCsv.add(localCsvOutput);
+				}
+				
+				SpatialCsvManager.mergeSortXY(totalCsvOutput, localCsv.toArray(new String[localCsv.size()]));
+				
+				for(String csv : localCsv){
+					new File(csv).delete();
+				}
+				
+			}
+			
+			
 		
-		SpatialCsvManager.mergeXY(totalCsvOutput, csvOutputs.toArray(new String[csvOutputs.size()]), "X", "Y", entete);
-		
-		for(String csvOut : csvOutputs){
-			new File(csvOut).delete();
-			Tool.copy(csvOut.replace(".csv", "_header.txt"), totalCsvOutput.replace(".csv", "_header.txt"));
-			new File(csvOut.replace(".csv", "_header.txt")).delete();
 		}
 		
 		totalMetrics = null;
